Move exp app pkg to cmd internal dir
