+++ exp/pm/procmgmt/cmd/svc/main.go
@@ -12,8 +12,8 @@ import (
 	"time"
 
 	"github.com/ActiveState/cli/exp/pm/internal/proccomm"
-	"github.com/ActiveState/cli/exp/pm/internal/serve"
 	"github.com/ActiveState/cli/exp/pm/internal/socket"
+	"github.com/ActiveState/cli/exp/pm/procmgmt/internal/serve"
 )
 
 func main() {
